move enum out of pages
